+++ src/components/Nav/ExternalNavItem.js
@@ -1,0 +1,9 @@
+export default function ExternalNavItem ({ children, ...restProps }) {
+  return (
+    <a {...restProps}>
+      <span>
+        {children}
+      </span>
+    </a>
+  )
+}